Add a script to write an Elm module with the current API URL
This needs to be written by a script after Terraform has provisioned everything,
because until then we don't know what the invocation URL of the API is. It would
be possible to use a custom (standard) domain for the API, but that would
require hardcoding of some kind. Easier just to keep it dynamic.